+++ ksvg/src/main/kotlin/hu/oandras/ksvg/render/RenderNode.kt
@@ -750,7 +750,7 @@ internal class FeComponentTransferRenderNode(
 
     // Lazily built [A,R,G,B] 256-entry LUTs; null until first use. The tables
     // depend only on build-time transfer functions, so they are computed once.
-    @JvmField var lutTables: Array<ByteArray>? = null
+    @JvmField var lutTables: Array<IntArray>? = null
 
     @JvmField var gpuLutBitmap: Bitmap? = null
 }
